worker/requirements.txt
pymongo==4.16.0
python-dotenv==1.2.2
redis==7.4.0